TECH-2322 expose es6 modules: rename data_table -> local_data_table
