+++ src/plugins/data_table/local_data_table.js
